Refactoring the column auto increment logic to a separated class.
